+++ packages/editor-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pptx-glimpse/editor-core
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [f0136a9]
+- Updated dependencies [c57532b]
+- Updated dependencies [f627f71]
+- Updated dependencies [3d61817]
+- Updated dependencies [020f949]
+- Updated dependencies [69ae720]
+- Updated dependencies [8904a5c]
+  - @pptx-glimpse/document@0.2.0
+
 ## 0.0.1
 
 ### Patch Changes